+++ integration-tests/generate_data_multi_command.py
@@ -1,9 +1,8 @@
-import os
+import subprocess
 import pandas as pd
 
-
 ### Input fa
-os.system('SemiBin generate_data_multi -i test/multi_samples_data/input_multi.fasta -o output_multi_fa -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :')
+subprocess.check_call('SemiBin generate_sequence_features_multi -i test/multi_samples_data/input_multi.fasta -o output_multi_fa -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :', shell=True)
 
 for i in range(10):
     data = pd.read_csv('output_multi_fa/samples/S{}/data.csv'.format(i + 1),
@@ -15,7 +14,7 @@ for i in range(10):
 
 
 ### Input .gz
-os.system('SemiBin generate_data_multi -i test/multi_samples_data/input_multi.fasta.gz -o output_multi_gz -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :')
+subprocess.check_call('SemiBin generate_sequence_features_multi -i test/multi_samples_data/input_multi.fasta.gz -o output_multi_gz -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :', shell=True)
 
 for i in range(10):
     data = pd.read_csv('output_multi_gz/samples/S{}/data.csv'.format(i + 1),
@@ -26,7 +25,7 @@ for i in range(10):
     assert data_split.shape == (40, 146)
 
 ### Input .bz2
-os.system('SemiBin generate_data_multi -i test/multi_samples_data/input_multi.fasta.bz2 -o output_multi_bz2 -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :')
+subprocess.check_call('SemiBin generate_sequence_features_multi -i test/multi_samples_data/input_multi.fasta.bz2 -o output_multi_bz2 -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :', shell=True)
 
 for i in range(10):
     data = pd.read_csv('output_multi_bz2/samples/S{}/data.csv'.format(i + 1),
@@ -37,7 +36,7 @@ for i in range(10):
     assert data_split.shape == (40, 146)
 
 ### Input .xz
-os.system('SemiBin generate_data_multi -i test/multi_samples_data/input_multi.fasta.xz -o output_multi_xz -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :')
+subprocess.check_call('SemiBin generate_sequence_features_multi -i test/multi_samples_data/input_multi.fasta.xz -o output_multi_xz -m 2500 --ratio 0.05 --ml-threshold 4000 -p 1 -b test/multi_samples_data/input_multi_sorted*.bam -s :', shell=True)
 
 for i in range(10):
     data = pd.read_csv('output_multi_xz/samples/S{}/data.csv'.format(i + 1),